requirements.txt
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
distro==1.9.0
Django==5.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-decouple==3.8
requests==2.32.4
sniffio==1.3.1
sqlparse==0.5.3
stripe==12.2.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
